chore: centralize translation helpers
