+++ api-server/services/translationsExport.js
@@ -2,53 +2,11 @@ import fs from 'fs';
 import path from 'path';
 import { getConfigPathSync } from '../utils/configPaths.js';
 import { slugify } from '../utils/slugify.js';
-
-function sortObj(o) {
-  return Object.keys(o)
-    .sort()
-    .reduce((acc, k) => ((acc[k] = o[k]), acc), {});
-}
-
-function collectPhrasesFromPages(dir) {
-  const files = [];
-  function walk(d) {
-    for (const entry of fs.readdirSync(d, { withFileTypes: true })) {
-      const full = path.join(d, entry.name);
-      if (entry.isDirectory()) walk(full);
-      else if (/\.(jsx?|tsx?)$/.test(entry.name)) files.push(full);
-    }
-  }
-  walk(dir);
-  const regex = /\bt\(\s*['"]([^'"]+)['"]\s*(?:,\s*['"]([^'"]+)['"])?\s*\)/g;
-  const pairs = [];
-  for (const file of files) {
-    const content = fs.readFileSync(file, 'utf8');
-    let match;
-    while ((match = regex.exec(content))) {
-      pairs.push({ key: match[1], text: match[2] || match[1] });
-    }
-  }
-  return pairs;
-}
-
-async function fetchModules() {
-  try {
-    const db = await import('../../db/index.js');
-    try {
-      const [rows] = await db.pool.query(
-        'SELECT module_key AS moduleKey, label FROM modules',
-      );
-      await db.pool.end();
-      return rows.map((r) => ({ moduleKey: r.moduleKey, label: r.label }));
-    } catch (err) {
-      try {
-        await db.pool.end();
-      } catch {}
-    }
-  } catch {}
-  const fallback = await import('../../db/defaultModules.js');
-  return fallback.default.map(({ moduleKey, label }) => ({ moduleKey, label }));
-}
+import {
+  collectPhrasesFromPages,
+  fetchModules,
+  sortObj,
+} from '../utils/translationHelpers.js';
 
 export async function exportTranslations(companyId = 0) {
   const { path: headerMappingsPath } = getConfigPathSync(
